+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- AI feature highlight row on onboarding features page
+
 ## [0.5.0] - 2026-02-19
 
 ### Changed